requirements.txt
python-dotenv
uuid

langchain
langchain-community
langchain-openai
langchain-huggingface

sentence-transformers
huggingface-hub



openai>=1.0.0

tqdm
pandas
numpy

streamlit
chromadb
faiss-cpu